chore: move to utils.ts
